+++ src/modules/tools/measure/store/indexMeasure.js
@@ -2,7 +2,7 @@ import getters from "./gettersMeasure";
 import mutations from "./mutationsMeasure";
 import actions from "./actionsMeasure";
 import state from "./stateMeasure";
-import "../typedefs";
+import "../util/typedefs";
 
 export default {
     namespaced: true,